+++ src/main/java/redstonedubstep/mods/vanishmod/VanishEventListener.java
@@ -64,19 +64,19 @@ public class VanishEventListener {
 
 	@SubscribeEvent
 	public static void onInteractBlock(RightClickBlock event) {
-		if (event.getPlayer() instanceof ServerPlayer player && player.gameMode.getGameModeForPlayer() != GameType.SPECTATOR)
+		if (VanishConfig.CONFIG.indirectSoundSuppression.get() && event.getPlayer() instanceof ServerPlayer player && player.gameMode.getGameModeForPlayer() != GameType.SPECTATOR)
 			SoundSuppressionHelper.updateBlockHitResult(player, event.getHitVec());
 	}
 
 	@SubscribeEvent
 	public static void onInteractEntity(EntityInteract event) {
-		if (event.getPlayer() instanceof ServerPlayer player && player.gameMode.getGameModeForPlayer() != GameType.SPECTATOR)
+		if (VanishConfig.CONFIG.indirectSoundSuppression.get() && event.getPlayer() instanceof ServerPlayer player && player.gameMode.getGameModeForPlayer() != GameType.SPECTATOR)
 			SoundSuppressionHelper.updateEntityHitResult(player, event.getTarget());
 	}
 
 	@SubscribeEvent
 	public static void onAttackEntity(AttackEntityEvent event) {
-		if (event.getPlayer() instanceof ServerPlayer player && player.gameMode.getGameModeForPlayer() != GameType.SPECTATOR)
+		if (VanishConfig.CONFIG.indirectSoundSuppression.get() && event.getPlayer() instanceof ServerPlayer player && player.gameMode.getGameModeForPlayer() != GameType.SPECTATOR)
 			SoundSuppressionHelper.updateEntityHitResult(player, event.getTarget());
 	}