+++ src/main/java/redstonedubstep/mods/vanishmod/mixin/sound/MixinEntity.java
@@ -9,6 +9,7 @@ import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.MoverType;
 import net.minecraft.world.phys.Vec3;
+import redstonedubstep.mods.vanishmod.VanishConfig;
 import redstonedubstep.mods.vanishmod.misc.SoundSuppressionHelper;
 
 @Mixin(Entity.class)
@@ -16,7 +17,7 @@ public class MixinEntity {
 	//Invalidates the hit results of a vanished player if its position changes, because then their crosshair is most likely on a different block
 	@Inject(method = "move", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Entity;setPos(DDD)V", ordinal = 1))
 	private void onActualMove(MoverType type, Vec3 pos, CallbackInfo callbackInfo) {
-		if ((Object)this instanceof ServerPlayer player)
+		if (VanishConfig.CONFIG.indirectSoundSuppression.get() && (Object)this instanceof ServerPlayer player)
 			SoundSuppressionHelper.invalidateHitResults(player);
 	}
 }